code cleanup: various nits

+++ palmod/Game/DescTree.h
@@ -27,7 +27,7 @@ struct sDescTreeNode
     // Solely used within CGameClassByDir
     // Allows for reusing units from loose ROM sets by specifying their offset location
     // within the new monolithic ROM composition
-    size_t nROMShift = 0;
+    int nROMShift = 0;
 };
 
 class CDescTree
