+++ palmod/Game/Game_JOJOS_A.cpp
@@ -217,7 +217,7 @@ void CGame_JOJOS_A::LoadSpecificPaletteData(uint32_t nUnitId, uint32_t nPalId)
         }
 
         m_nCurrentPaletteROMLocation = nOffset;
-        m_nCurrentPaletteSizeInColors = cbPaletteSizeOnDisc / m_nSizeOfColorsInBytes;
+        m_nCurrentPaletteSizeInColors = static_cast<uint16_t>(cbPaletteSizeOnDisc / m_nSizeOfColorsInBytes);
     }
     else //Extra Palettes
     {
